Properly support multiple instances of this bot (#18)
Previously, it was hardcoded to attempt to attach to the zeroth
Guild identified; this caused problems if you used the same
Discord Application instance to connect to multiple Guilds.  This
is now fixed and it properly looks at the connected Guilds to match
against the ID in the `.env` file.